Migrate ToggleGroup from /docs/components/toggle-group to /components/toggle-group
