+++ site/ui/renderer.tsx
@@ -119,7 +119,7 @@ const menuEntries: SidebarEntry[] = [
       { title: 'Textarea', href: '/components/textarea' },
       { title: 'Toast', href: '/docs/components/toast' },
       { title: 'Toggle', href: '/components/toggle' },
-      { title: 'Toggle Group', href: '/docs/components/toggle-group' },
+      { title: 'Toggle Group', href: '/components/toggle-group' },
       { title: 'Tooltip', href: '/docs/components/tooltip' },
     ],
   },
